# Allow additionally parsing \r newlines in PEM files