… tests/data/mixed-line-endings.crt