@ctz 2023-11-08
I was thinking this might be a regression from ff11ce1b6b658d498e80515bab6843ed71db938f, but actually std's `read_line` is also obsessed with `\n`: https://doc.rust-lang.org/stable/src/std/io/mod.rs.html#2235-2240 🤨