@complexspaces 2023-11-08
That was my conclusion as well. You have to go out of your way to support other line endings in most functions, with the exception of [str::lines](https://doc.rust-lang.org/std/primitive.str.html#method.lines) which supports `CRLF` as well.